requirements.txt
flake8==3.9.2
flake8-docstrings==1.6.0
pytest==6.2.5
python-dotenv==0.20.0
python-telegram-bot==13.13
requests==2.26.0